Cleanup, made sure documentation and comments are consistent with changes made. Made seperate .c file for driver.
