+++ TOVola/README
@@ -27,7 +27,7 @@ RK4(5) and DP7(8). GSL allows for more methods, both adaptive and non-adaptive,
 but the two I listed felt sufficient enough.
 
 All these choices can be made in the parfile. No touching the src code is
-necessary (as it should). For examples of using each EOS type, see the par directory.
+necessary (as it should be). For examples of using each EOS type, see the par directory.
 
 ==========
 3. Output
@@ -44,12 +44,12 @@ constraint tested using Baikal.
 
 Simple Polytrope: 
 Raw data sits within 10^-12/-13 in relative error with NRpy
-Constraint violation goes down with higher resolved grids
+Constraint violation tested with Baikal
 
 Piecewise Polytrope:
 Raw data sits within 10^-12/-13 in relative error with NRpy
-Constraint violation goes down with higher resolved grids
+Constraint violation tested with Baikal
 
 Tabulated EOS:
 Raw data sits within 10^-8 in relative error with NRpy (Mostly from EOS table interpolation)
-Constraint violation goes down with higher resolved grids (to interpolation error)
+Constraint violation tested with Baikal
