+++ TOVola/interface.ccl
@@ -2,5 +2,3 @@ implements: TOVola
 inherits: GRHayLib, ADMBase, HydroBase, Grid
 
 public:
-
-# nothing for now, at the moment no other thorn needs to see this one.
