+++ TOVola/par/Tabulated.par
@@ -27,7 +27,7 @@ TmunuBase::prolongation_type     = none
 #---------TOVola---------
 TOVola::TOVola_Interpolation_Stencil = 8
 TOVola::TOVola_Tin = 1.0e-2
-TOVola::TOVola_ODE_method = "ARKF"
+TOVola::TOVola_ODE_method = "ADP8"
 TOVola::TOVola_central_baryon_density = 1.58e-3
 TOVola::TOVola_error_limit = 1.0e-8
 TOVola::TOVola_EOS_type = "Tabulated"
