+++ TOVola/param.ccl
@@ -77,7 +77,7 @@ CCTK_INT TOVola_size "Maximum number of steps"
 } 10000000
 
 #Only allowing these two methods right now. I can add a few more if there is a demand.
-STRING TOVola_ODE_method "What type of method are we using (More Methods can be added if there is demand)"
+STRING TOVola_ODE_method "What type of ODE method are we using (More Methods can be added if there is demand)"
 {
 	"ARKF" :: "Adaptive Runge-Kutta-Fehlberg (RK4(5))"
 	"ADP8" :: "Adaptive Dormand-Prince Eigth Order (DP7(8))"
@@ -103,7 +103,7 @@ CCTK_REAL TOVola_absolute_min_step "Smallest possible step size"
 #################################################################################################################
 #Populate Timelevels
 #################################################################################################################
-CCTK_INT TOVola_TOV_Populate_Timelevels "Populate that amount of timelevels (Based on original ET TOVSolver)" STEERABLE=always
+CCTK_INT TOVola_TOV_Populate_Timelevels "Populate that amount of timelevels" STEERABLE=always
 {
   1:3 :: "1 (default) to 3"
 } 1
