+++ TOVola/src/TOVola_driver.c
@@ -1,0 +1,351 @@
+#include "GRHayLib.h"
+#include "TOVola_interp.h"
+#include "TOVola_defines.h"
+#include "TOVola_solve.h"
+#include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <math.h>
+#include <gsl/gsl_odeiv2.h>
+#include <gsl/gsl_errno.h>
+
+#include <cctk.h>
+#include <cctk_Arguments.h>
+#include <cctk_Parameters.h>
+
+/********************************
+// TOVola is a TOV solver inspired by NRPy designed specifically for use in the Einstein Toolkit.
+// It integrates the TOV equations for a solution to static, spherically symmetric stars (Neutron stars in specific), and interpollates the solution the the ET grid.
+//
+// MOST IMPORTANTLY, it handles multiple types of EOS in the solution: Simple Polytrope, Piecewise Polytrope, and Tabulated EOS. An upgrade from the original ET TOVSolver thorn.
+//
+// Simple Polytrope:
+//      -- The Simple Polytrope is a special version of the Piecewise Polytrope with one EOS region relating Pressure to baryon density (P=K*rho_baryon^Gamma), where K and Gamma are parameters to be input
+//      -- Only one K and Gamma are necessary to run this implementation, and a Gamma_Thermal defaulted to 2.0
+//      -- Hamiltonian Constraint Violation Validated with Baikal
+// Piecewise Polytrope:
+//      -- The more general verion of the simple polytrope where you can choose the number of regions for your polytrope, each region denoted by (P=K_i*rho_baryon^Gamma_i)
+//      -- You can get your necessary parameters by following the method outlined in Read et. al. (2008)
+//      -- Since we follow Read's method, you only need to input one K value (The GRHayL library will handle the rest), and all the Gammas and rho_baryon boundary points for the number of specified regions.
+//      -- Hamiltonian Constraint Violation Validated with Baikal
+// Tabulated EOS:
+//      -- The tabulated solver reads in a EOS table from your computer.
+//      -- Once located, GRHayL slices the table for beta_equilibirum and then uses an interpollator to find the necessary densities through the calculations.
+//      -- Hamiltonian Constraint Violation Validated with Baikal
+//
+// Users who wish to interface with this program need only create a parfile for a simulation. Output variables will be adjusted for use in ADMbase and Hydrobase as initial data.
+// IllinoisGRMHD is used to set Tmunu from the data and Baikal tests the constraint Violation.
+// 
+// Gallery examples are currently in the works, using GRHydro for evolution. However, we have also used TOVola with GRHayL evolutions for a paper currently in prep.
+// 
+// David Boyer (10/08/24)
+********************************/
+
+#define ODE_SOLVER_DIM 4
+#define TOVOLA_PRESSURE 0
+#define TOVOLA_NU 1
+#define TOVOLA_MASS 2
+#define TOVOLA_R_ISO 3
+#define NEGATIVE_R_INTERP_BUFFER 11
+
+//Helpful defines for later.
+#define velx (&vel[0*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define vely (&vel[1*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define velz (&vel[2*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+//For timelevel population
+#define velx_p (&vel_p[0*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define vely_p (&vel_p[1*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define velz_p (&vel_p[2*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define velx_p_p (&vel_p_p[0*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define vely_p_p (&vel_p_p[1*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+#define velz_p_p (&vel_p_p[2*cctk_lsh[0]*cctk_lsh[1]*cctk_lsh[2]])
+
+
+
+//Perform the TOV integration using GSL
+void TOVola_Solve_and_Interp(CCTK_ARGUMENTS){
+
+  DECLARE_CCTK_PARAMETERS;
+  DECLARE_CCTK_ARGUMENTS;
+  TOVola_ID_persist_struct TOVola_ID_persist_tmp;  // allocates memory for the pointer below.
+  TOVola_ID_persist_struct *restrict TOVola_ID_persist = &TOVola_ID_persist_tmp;
+  CCTK_REAL current_position = 0;
+
+  /* Set up ODE system and driver */
+  TOVola_data_struct TOVdata_tmp; // allocates memory for the pointer below.
+  TOVola_data_struct *restrict TOVdata = &TOVdata_tmp;
+  gsl_odeiv2_system system;
+  gsl_odeiv2_driver *driver;
+
+  //Checking and setting EOS
+  if(CCTK_EQUALS("Simple",TOVola_EOS_type)){
+    CCTK_INFO("Simple Polytrope");
+    TOVdata->eos_type = 0;
+    if(ghl_eos->neos!=1){
+      CCTK_INFO("Error: Too many regions for the simple polytrope.");
+      CCTK_INFO("Check your value for neos, or use a piecewise polytrope");
+      CCTK_ERROR("Shutting down due to error...");}       
+    }
+  else if(CCTK_EQUALS("Piecewise",TOVola_EOS_type)){
+    CCTK_INFO("Piecewise Polytrope");
+    TOVdata->eos_type=1;}
+  else if(CCTK_EQUALS("Tabulated",TOVola_EOS_type)){
+    CCTK_INFO("Tabulated EOS");
+    TOVdata->eos_type=2;
+    ghl_tabulated_compute_Ye_P_eps_of_rho_beq_constant_T(TOVola_Tin, ghl_eos);
+  }
+  else{
+    CCTK_INFO("ERROR: Invalid EOS type. Must be either 'Simple', 'Piecewise', or 'Tabulated'");
+    CCTK_ERROR("Shutting down due to error...");}
+
+  //Initialize other TOVdata member variables
+  TOVdata->numpoints_actually_saved = 0;
+  TOVdata->error_limit = TOVola_error_limit;
+  TOVdata->initial_ode_step_size = TOVola_initial_ode_step_size;
+  TOVdata->absolute_max_step = TOVola_absolute_max_step;
+  TOVdata->absolute_min_step = TOVola_absolute_min_step;
+  TOVdata->central_baryon_density = TOVola_central_baryon_density;
+  TOVdata->ghl_eos = ghl_eos;
+  if (setup_ode_system(TOVola_ODE_method, &system, &driver, TOVdata) != 0) {
+    CCTK_ERROR("Failed to set up ODE system.");
+  }
+  
+  CCTK_INFO("Starting TOV Integration using GSL for TOVola...");
+  /* Initialize ODE variables */
+  CCTK_REAL TOVola_eq[ODE_SOLVER_DIM];
+  CCTK_REAL c[2];
+  TOVola_get_initial_condition(TOVola_eq, TOVdata);
+  TOVola_assign_constants(c, TOVdata);
+
+  /* Initial memory allocation */
+  TOVdata->numels_alloced_TOV_arr = 1024;
+  if (initialize_tovola_data(TOVdata) != 0) {
+    gsl_odeiv2_driver_free(driver);
+    CCTK_ERROR("Failed to initialize TOVola_data_struct.");
+  }
+
+  /* Integration loop */
+  TOVdata->r_lengthscale = TOVola_initial_ode_step_size; // initialize dr to a crazy small value in double precision.
+  for (int i = 0; i < TOVola_size; i++) {
+    CCTK_REAL dr = 0.01 * TOVdata->r_lengthscale;
+    if (TOVdata->rho_baryon < 0.05 * TOVola_central_baryon_density) {
+      // To get a super-accurate mass, reduce the dr sampling near the surface of the star.
+      dr = 1e-6 * TOVdata->r_lengthscale;
+    }
+    /* Exception handling */
+    TOVola_exception_handler(current_position, TOVola_eq);
+
+    /* Apply ODE step */
+    int status = gsl_odeiv2_driver_apply(driver, &current_position, current_position + dr, TOVola_eq);
+    if (status != GSL_SUCCESS) {
+      CCTK_VINFO("GSL ODE solver failed with status %d.", status);
+      gsl_odeiv2_driver_free(driver);
+      CCTK_ERROR("Shutting down due to error");
+    };
+
+    /* Post-step exception handling */
+    TOVola_exception_handler(current_position, TOVola_eq);
+
+    /* Evaluate densities */
+    TOVola_evaluate_rho_and_eps(current_position, TOVola_eq, TOVdata);
+    TOVola_assign_constants(c, TOVdata);
+
+    /* Check if reallocation is needed */
+    if (TOVdata->numpoints_actually_saved >= TOVdata->numels_alloced_TOV_arr) {
+      // Update arr_size instead of modifying the macro
+      const int new_arr_size = 1.5 * TOVdata->numels_alloced_TOV_arr;
+      TOVdata->numels_alloced_TOV_arr = new_arr_size;
+      TOVdata->rSchw_arr = realloc(TOVdata->rSchw_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->rho_energy_arr = realloc(TOVdata->rho_energy_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->rho_baryon_arr = realloc(TOVdata->rho_baryon_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->P_arr = realloc(TOVdata->P_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->M_arr = realloc(TOVdata->M_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->nu_arr = realloc(TOVdata->nu_arr, sizeof(CCTK_REAL) * new_arr_size);
+      TOVdata->Iso_r_arr = realloc(TOVdata->Iso_r_arr, sizeof(CCTK_REAL) * new_arr_size);
+
+      if (!TOVdata->rSchw_arr || !TOVdata->rho_energy_arr || !TOVdata->rho_baryon_arr || !TOVdata->P_arr || !TOVdata->M_arr || !TOVdata->nu_arr ||
+          !TOVdata->Iso_r_arr) {
+        CCTK_ERROR("Memory reallocation failed during integration.\n");
+        gsl_odeiv2_driver_free(driver);
+      }
+    }
+
+    /* Store data */
+    TOVdata->rSchw_arr[TOVdata->numpoints_actually_saved] = current_position;
+    TOVdata->rho_energy_arr[TOVdata->numpoints_actually_saved] = c[0];
+    TOVdata->rho_baryon_arr[TOVdata->numpoints_actually_saved] = c[1];
+    TOVdata->P_arr[TOVdata->numpoints_actually_saved] = TOVola_eq[TOVOLA_PRESSURE];
+    TOVdata->M_arr[TOVdata->numpoints_actually_saved] = TOVola_eq[TOVOLA_MASS];
+    TOVdata->nu_arr[TOVdata->numpoints_actually_saved] = TOVola_eq[TOVOLA_NU];
+    TOVdata->Iso_r_arr[TOVdata->numpoints_actually_saved] = TOVola_eq[TOVOLA_R_ISO];
+    TOVdata->numpoints_actually_saved++;
+
+    /* Termination condition */
+    if (TOVola_do_we_terminate(current_position, TOVola_eq, TOVdata)) {
+      CCTK_VINFO("Finished Integration at position %.6e with Mass %.14e", current_position, TOVola_eq[TOVOLA_MASS]);
+      break;
+    }
+  }
+
+  /* Cleanup */
+  gsl_odeiv2_driver_free(driver);
+  CCTK_INFO("ODE Solver using GSL for TOVola Shutting Down...");
+
+  // Data in TOVdata->*_arr are stored at r=TOVdata->initial_ode_step_size > 0 up to the stellar surface.
+  // However, we may need data at r=0, which would require extrapolation.
+  // To prevent that, we copy INTERP_BUFFER data points from r>0 to r<0 so that we can always interpolate.
+  CCTK_REAL *restrict tmp = malloc(sizeof(CCTK_REAL) * (TOVdata->numpoints_actually_saved + NEGATIVE_R_INTERP_BUFFER));
+
+  extend_to_negative_r(TOVdata->rSchw_arr, -1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->rho_energy_arr, +1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->rho_baryon_arr, +1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->P_arr, +1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->M_arr, +1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->nu_arr, +1.0, tmp, TOVdata);
+  extend_to_negative_r(TOVdata->Iso_r_arr, -1.0, tmp, TOVdata);
+
+  free(tmp);
+  TOVdata->numpoints_actually_saved += NEGATIVE_R_INTERP_BUFFER;
+
+  /* Allocate and populate TOVola_ID_persist_struct arrays */
+  TOVola_ID_persist->r_Schw_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->rho_energy_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->rho_baryon_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->P_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->M_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->expnu_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->exp4phi_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+  TOVola_ID_persist->r_iso_arr = (CCTK_REAL *restrict)malloc(sizeof(CCTK_REAL) * TOVdata->numpoints_actually_saved);
+
+  if (!TOVola_ID_persist->r_Schw_arr || !TOVola_ID_persist->rho_energy_arr || !TOVola_ID_persist->rho_baryon_arr || !TOVola_ID_persist->P_arr || !TOVola_ID_persist->M_arr ||
+      !TOVola_ID_persist->expnu_arr || !TOVola_ID_persist->exp4phi_arr || !TOVola_ID_persist->r_iso_arr) {
+    free_tovola_data(TOVdata);
+    CCTK_ERROR("Memory allocation failed for TOVola_ID_persist_struct arrays.\n");
+  }
+
+  TOVola_ID_persist->numpoints_arr = TOVdata->numpoints_actually_saved;
+
+  /* Normalize and set data */
+  TOVola_Normalize_and_set_data_integrated(TOVdata, TOVola_ID_persist->r_Schw_arr, TOVola_ID_persist->rho_energy_arr, TOVola_ID_persist->rho_baryon_arr, TOVola_ID_persist->P_arr,
+                                           TOVola_ID_persist->M_arr, TOVola_ID_persist->expnu_arr, TOVola_ID_persist->exp4phi_arr, TOVola_ID_persist->r_iso_arr);
+
+  /* Free raw data as it's no longer needed */
+  free_tovola_data(TOVdata);
+
+  /* Now to interp, and finalize the grid. */
+  CCTK_REAL TOVola_Rbar = TOVola_ID_persist->r_iso_arr[TOVola_ID_persist->numpoints_arr-1]; 
+  CCTK_REAL TOVola_Mass = TOVola_ID_persist->M_arr[TOVola_ID_persist->numpoints_arr-1];
+  
+  
+  //Now for the actual grid placements. Go over all grid points
+  CCTK_INFO("TOVola Beginning Grid Placements...");
+#pragma omp parallel for
+  for(int i=0; i<cctk_lsh[0]; i++){ 
+  	for(int j=0; j<cctk_lsh[1]; j++){ 
+  		for(int k=0; k<cctk_lsh[2]; k++){ 
+  			int i3d=CCTK_GFINDEX3D(cctkGH,i,j,k); //3D index
+  			CCTK_REAL TOVola_r_iso = sqrt((x[i3d]*x[i3d])+(y[i3d]*y[i3d])+(z[i3d]*z[i3d])); //magnitude of r on the grid
+  			CCTK_REAL TOVola_rho_energy, TOVola_rho_baryon, TOVola_P, TOVola_M, TOVola_expnu, TOVola_exp4phi; //Declare TOV quantities
+  			if (TOVola_r_iso < TOVola_Rbar){ //If we are INSIDE the star, we need to interpollate the data to the grid.
+  				TOVola_TOV_interpolate_1D(TOVola_r_iso, TOVola_Interpolation_Stencil, TOVola_Max_Interpolation_Stencil,
+                                      TOVola_ID_persist->numpoints_arr, TOVola_ID_persist->r_Schw_arr,
+                                      TOVola_ID_persist->rho_energy_arr, TOVola_ID_persist->rho_baryon_arr, TOVola_ID_persist->P_arr,
+                                      TOVola_ID_persist->M_arr, TOVola_ID_persist->expnu_arr, TOVola_ID_persist->exp4phi_arr,
+                                      TOVola_ID_persist->r_iso_arr, &TOVola_rho_energy, &TOVola_rho_baryon, &TOVola_P,
+                                      &TOVola_M, &TOVola_expnu, &TOVola_exp4phi);
+  				rho[i3d] = TOVola_rho_baryon;
+				press[i3d] = TOVola_P;
+				// tiny number prevents 0/0.
+				eps[i3d] = (TOVola_rho_energy / (TOVola_rho_baryon+1e-30)) - 1.0;
+				if (eps[i3d]<0){eps[i3d]=0.0;}
+				alp[i3d] = pow(TOVola_expnu,0.5);//This is the lapse
+				gxx[i3d] = TOVola_exp4phi;//This is the values for the metric in the coordinates we chose.
+				gyy[i3d] = gxx[i3d];
+				gzz[i3d] = gxx[i3d];}
+			else { //If we are OUTSIDE the star, we need to calculate the grid functions directly. Thank you, Schwarzchild.
+				CCTK_REAL TOVola_rSchw_outside = (TOVola_r_iso+TOVola_Mass) + TOVola_Mass*TOVola_Mass/(4.0*TOVola_r_iso);//Need to know what rSchw is at our current grid location.
+				rho[i3d] = 0.0;
+				press[i3d] = 0.0;
+				eps[i3d] = 0.0;
+				alp[i3d] = pow(1-2*TOVola_Mass/TOVola_rSchw_outside,0.5); //Goes to Schwarschild
+				gxx[i3d] = pow((TOVola_rSchw_outside/TOVola_r_iso),2.0);
+				gyy[i3d] = gxx[i3d];
+				gzz[i3d] = gxx[i3d];
+				}
+			
+			betax[i3d] = 0.0;
+			betay[i3d] = 0.0;
+			betaz[i3d] = 0.0;
+			gxy[i3d] = 0.0;
+			gxz[i3d] = 0.0;
+			gyz[i3d] = 0.0;
+			//Curvature is zero for our slice.
+			kxx[i3d] = 0.0;
+			kyy[i3d] = 0.0;
+			kzz[i3d] = 0.0;
+			kxy[i3d] = 0.0;
+			kxz[i3d] = 0.0;
+			kyz[i3d] = 0.0;
+			//velocities are zero: It is a static solution.
+			velx[i3d] = 0.0;
+			vely[i3d] = 0.0;
+			velz[i3d] = 0.0;
+	        	w_lorentz[i3d] = 1.0;
+  		}
+  	}
+  }
+  			
+  CCTK_INFO("Grid Placement Successful!");
+  CCTK_INFO("Populating Time Levels...");
+
+  //This is to populate time levels.
+  //Luckily, this is a static solution, so the logic isn't too complicated.
+  //From the original TOVsolver in the toolkit.
+  int i3d = cctk_lsh[2]*cctk_lsh[1]*cctk_lsh[0];
+  switch(TOVola_TOV_Populate_Timelevels)
+  {
+    case 3:
+        TOVola_TOV_Copy(i3d, gxx_p_p,  gxx);
+        TOVola_TOV_Copy(i3d, gyy_p_p,  gyy);
+        TOVola_TOV_Copy(i3d, gzz_p_p,  gzz);
+        TOVola_TOV_Copy(i3d, gxy_p_p,  gxy);
+        TOVola_TOV_Copy(i3d, gxz_p_p,  gxz);
+        TOVola_TOV_Copy(i3d, gyz_p_p,  gyz);
+        TOVola_TOV_Copy(i3d, rho_p_p,  rho);
+        TOVola_TOV_Copy(i3d, eps_p_p,  eps);
+        TOVola_TOV_Copy(i3d, velx_p_p, velx);
+        TOVola_TOV_Copy(i3d, vely_p_p, vely);
+        TOVola_TOV_Copy(i3d, velz_p_p, velz);
+        TOVola_TOV_Copy(i3d, w_lorentz_p_p, w_lorentz);
+        // fall through
+    case 2:
+        TOVola_TOV_Copy(i3d, gxx_p,  gxx);
+        TOVola_TOV_Copy(i3d, gyy_p,  gyy);
+        TOVola_TOV_Copy(i3d, gzz_p,  gzz);
+        TOVola_TOV_Copy(i3d, gxy_p,  gxy);
+        TOVola_TOV_Copy(i3d, gxz_p,  gxz);
+        TOVola_TOV_Copy(i3d, gyz_p,  gyz);
+        TOVola_TOV_Copy(i3d, rho_p,  rho);
+        TOVola_TOV_Copy(i3d, eps_p,  eps);
+        TOVola_TOV_Copy(i3d, velx_p, velx);
+        TOVola_TOV_Copy(i3d, vely_p, vely);
+        TOVola_TOV_Copy(i3d, velz_p, velz);
+        TOVola_TOV_Copy(i3d, w_lorentz_p, w_lorentz);
+        // fall through
+    case 1:
+        break;
+    default:
+        CCTK_VWARN(CCTK_WARN_ABORT,
+                   "Unsupported number of TOVola_TOV_Populate_TimelevelsL: %d",
+                   (int)TOVola_TOV_Populate_Timelevels);
+        break;
+  }
+
+
+  CCTK_INFO("Population Complete!");
+  
+  free_ID_persist_data(TOVola_ID_persist);
+  
+  CCTK_INFO("Complete! Enjoy your initial data!");
+  CCTK_INFO("TOVola shutting down...");
+}
