+++ TOVola/src/make.code.defn
@@ -1,3 +1,3 @@
-SRCS = TOVola_solve.c
+SRCS = TOVola_driver.c
 
 SUBDIRS =
