+++ TOVola/test/Piecewise_test.par
@@ -25,15 +25,11 @@ TmunuBase::timelevels            =  1
 TmunuBase::prolongation_type     = none
 
 #---------TOVola---------
-TOVola::TOVola_step = 1.0e-5
-TOVola::TOVola_size = 100000000
 TOVola::TOVola_Interpolation_Stencil = 12
 TOVola::TOVola_ODE_method = "ADP8"
 TOVola::TOVola_central_baryon_density = 1.58e-3
 TOVola::TOVola_error_limit = 1.0e-8
 TOVola::TOVola_EOS_type = "Piecewise"
-TOVola::TOVola_absolute_max_step = 1.0e-1
-TOVola::TOVola_absolute_min_step = 1.0e-8
 
 #---------ADMBase---------
 ADMBase::initial_data            = "TOVola"
