+++ TOVola/test/Piecewise_test/HGF.x.asc
@@ -5,45 +5,45 @@
 0	2 24 24	0	-7.3333333333333 0 0	-nan
 0	3 24 24	0	-7 0 0	-nan
 0	4 24 24	0	-6.6666666666667 0 0	-nan
-0	5 24 24	0	-6.3333333333333 0 0	-8.0965018916798e-05
-0	6 24 24	0	-6 0 0	4.9801740139375e-05
-0	7 24 24	0	-5.6666666666667 0 0	-0.0014132591504823
-0	8 24 24	0	-5.3333333333333 0 0	0.00010072973929169
-0	9 24 24	0	-5 0 0	8.2979101114142e-06
-0	10 24 24	0	-4.6666666666667 0 0	1.0957640869252e-06
-0	11 24 24	0	-4.3333333333333 0 0	2.4059390579742e-06
-0	12 24 24	0	-4 0 0	2.7250002879176e-09
-0	13 24 24	0	-3.6666666666667 0 0	1.0429682954088e-07
-0	14 24 24	0	-3.3333333333333 0 0	3.2248498378629e-08
-0	15 24 24	0	-3 0 0	-1.4940778184092e-08
-0	16 24 24	0	-2.6666666666667 0 0	-5.7773455922061e-08
-0	17 24 24	0	-2.3333333333333 0 0	-1.0614864327796e-07
-0	18 24 24	0	-2 0 0	-1.645380413029e-07
-0	19 24 24	0	-1.6666666666667 0 0	-2.3320316282216e-07
-0	20 24 24	0	-1.3333333333333 0 0	-3.0837381288784e-07
-0	21 24 24	0	-1 0 0	-3.8266295752931e-07
-0	22 24 24	0	-0.66666666666667 0 0	-4.4633948680095e-07
-0	23 24 24	0	-0.33333333333333 0 0	-4.8951936730546e-07
-0	24 24 24	0	0 0 0	-5.0487831551593e-07
-0	25 24 24	0	0.33333333333333 0 0	-4.8951937316188e-07
-0	26 24 24	0	0.66666666666667 0 0	-4.4633946449935e-07
-0	27 24 24	0	1 0 0	-3.8266296902012e-07
-0	28 24 24	0	1.3333333333333 0 0	-3.0837383448168e-07
-0	29 24 24	0	1.6666666666667 0 0	-2.3320317241171e-07
-0	30 24 24	0	2 0 0	-1.6453804152494e-07
-0	31 24 24	0	2.3333333333333 0 0	-1.0614864229264e-07
-0	32 24 24	0	2.6666666666667 0 0	-5.7773423878249e-08
-0	33 24 24	0	3 0 0	-1.4940771661531e-08
-0	34 24 24	0	3.3333333333333 0 0	3.2248463871509e-08
-0	35 24 24	0	3.6666666666667 0 0	1.0429682929108e-07
-0	36 24 24	0	4 0 0	2.7250001977119e-09
-0	37 24 24	0	4.3333333333333 0 0	2.4059390471148e-06
-0	38 24 24	0	4.6666666666667 0 0	1.0957640597664e-06
-0	39 24 24	0	5 0 0	8.2979101116605e-06
-0	40 24 24	0	5.3333333333333 0 0	0.00010072973929077
-0	41 24 24	0	5.6666666666667 0 0	-0.0014132591504871
-0	42 24 24	0	6 0 0	4.9801740137094e-05
-0	43 24 24	0	6.3333333333333 0 0	-8.0965018890542e-05
+0	5 24 24	0	-6.3333333333333 0 0	-8.0985651460177e-05
+0	6 24 24	0	-6 0 0	5.0123420151424e-05
+0	7 24 24	0	-5.6666666666667 0 0	-0.0013712429671282
+0	8 24 24	0	-5.3333333333333 0 0	0.00010103144369296
+0	9 24 24	0	-5 0 0	8.2797565462769e-06
+0	10 24 24	0	-4.6666666666667 0 0	1.0963562451435e-06
+0	11 24 24	0	-4.3333333333333 0 0	2.3840556819046e-06
+0	12 24 24	0	-4 0 0	2.7566289598213e-09
+0	13 24 24	0	-3.6666666666667 0 0	1.0429277257085e-07
+0	14 24 24	0	-3.3333333333333 0 0	3.2248536299684e-08
+0	15 24 24	0	-3 0 0	-1.4941036755034e-08
+0	16 24 24	0	-2.6666666666667 0 0	-5.7773130640593e-08
+0	17 24 24	0	-2.3333333333333 0 0	-1.0614812939735e-07
+0	18 24 24	0	-2 0 0	-1.6453792103799e-07
+0	19 24 24	0	-1.6666666666667 0 0	-2.3320311066943e-07
+0	20 24 24	0	-1.3333333333333 0 0	-3.0837376213677e-07
+0	21 24 24	0	-1 0 0	-3.8266290941502e-07
+0	22 24 24	0	-0.66666666666667 0 0	-3.3804601513865e-07
+0	23 24 24	0	-0.33333333333333 0 0	-2.2019549399538e-06
+0	24 24 24	0	0 0 0	9.0615232549579e-06
+0	25 24 24	0	0.33333333333333 0 0	-2.201954913697e-06
+0	26 24 24	0	0.66666666666667 0 0	-3.380460134178e-07
+0	27 24 24	0	1 0 0	-3.8266290511291e-07
+0	28 24 24	0	1.3333333333333 0 0	-3.0837379483284e-07
+0	29 24 24	0	1.6666666666667 0 0	-2.3320310858777e-07
+0	30 24 24	0	2 0 0	-1.6453792028859e-07
+0	31 24 24	0	2.3333333333333 0 0	-1.0614818285459e-07
+0	32 24 24	0	2.6666666666667 0 0	-5.7773080833212e-08
+0	33 24 24	0	3 0 0	-1.4941042306149e-08
+0	34 24 24	0	3.3333333333333 0 0	3.2248520590028e-08
+0	35 24 24	0	3.6666666666667 0 0	1.0429280503793e-07
+0	36 24 24	0	4 0 0	2.7566128130152e-09
+0	37 24 24	0	4.3333333333333 0 0	2.3840557147187e-06
+0	38 24 24	0	4.6666666666667 0 0	1.0963562688746e-06
+0	39 24 24	0	5 0 0	8.2796388965779e-06
+0	40 24 24	0	5.3333333333333 0 0	0.00010103340098175
+0	41 24 24	0	5.6666666666667 0 0	-0.00137125539846
+0	42 24 24	0	6 0 0	5.0125515693565e-05
+0	43 24 24	0	6.3333333333333 0 0	-8.0985785924951e-05
 0	44 24 24	0	6.6666666666667 0 0	-nan
 0	45 24 24	0	7 0 0	-nan
 0	46 24 24	0	7.3333333333333 0 0	-nan
