+++ TOVola/test/Piecewise_test/alp.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-8 0 0	0.8254722683968
-0	1 24 24	0	-7.6666666666667 0 0	0.81857245696918
-0	2 24 24	0	-7.3333333333333 0 0	0.81110463279419
-0	3 24 24	0	-7 0 0	0.80299564412117
-0	4 24 24	0	-6.6666666666667 0 0	0.79415921461071
-0	5 24 24	0	-6.3333333333333 0 0	0.78449286165898
-0	6 24 24	0	-6 0 0	0.7738739281459
-0	7 24 24	0	-5.6666666666667 0 0	0.76218720931443
-0	8 24 24	0	-5.3333333333333 0 0	0.74964251188883
-0	9 24 24	0	-5 0 0	0.73653458937031
-0	10 24 24	0	-4.6666666666667 0 0	0.72305160388213
-0	11 24 24	0	-4.3333333333333 0 0	0.70936213909141
-0	12 24 24	0	-4 0 0	0.69563187378245
-0	13 24 24	0	-3.6666666666667 0 0	0.68203032219265
-0	14 24 24	0	-3.3333333333333 0 0	0.66873262260909
-0	15 24 24	0	-3 0 0	0.65591962493891
-0	16 24 24	0	-2.6666666666667 0 0	0.64377655871939
-0	17 24 24	0	-2.3333333333333 0 0	0.63249037799674
-0	18 24 24	0	-2 0 0	0.62224591745296
-0	19 24 24	0	-1.6666666666667 0 0	0.61322102228002
-0	20 24 24	0	-1.3333333333333 0 0	0.60558089747735
-0	21 24 24	0	-1 0 0	0.59947202976392
-0	22 24 24	0	-0.66666666666667 0 0	0.59501613838967
-0	23 24 24	0	-0.33333333333333 0 0	0.59230468004759
-0	24 24 24	0	0 0 0	0.59139444102415
-0	25 24 24	0	0.33333333333333 0 0	0.59230468004759
-0	26 24 24	0	0.66666666666667 0 0	0.59501613838967
-0	27 24 24	0	1 0 0	0.59947202976392
-0	28 24 24	0	1.3333333333333 0 0	0.60558089747735
-0	29 24 24	0	1.6666666666667 0 0	0.61322102228002
-0	30 24 24	0	2 0 0	0.62224591745296
-0	31 24 24	0	2.3333333333333 0 0	0.63249037799674
-0	32 24 24	0	2.6666666666667 0 0	0.64377655871939
-0	33 24 24	0	3 0 0	0.65591962493891
-0	34 24 24	0	3.3333333333333 0 0	0.66873262260909
-0	35 24 24	0	3.6666666666667 0 0	0.68203032219265
-0	36 24 24	0	4 0 0	0.69563187378245
-0	37 24 24	0	4.3333333333333 0 0	0.70936213909141
-0	38 24 24	0	4.6666666666667 0 0	0.72305160388213
-0	39 24 24	0	5 0 0	0.73653458937031
-0	40 24 24	0	5.3333333333333 0 0	0.74964251188883
-0	41 24 24	0	5.6666666666667 0 0	0.76218720931443
-0	42 24 24	0	6 0 0	0.7738739281459
-0	43 24 24	0	6.3333333333333 0 0	0.78449286165898
-0	44 24 24	0	6.6666666666667 0 0	0.79415921461071
-0	45 24 24	0	7 0 0	0.80299564412117
-0	46 24 24	0	7.3333333333333 0 0	0.81110463279419
-0	47 24 24	0	7.6666666666667 0 0	0.81857245696918
-0	48 24 24	0	8 0 0	0.8254722683968
+0	0 24 24	0	-8 0 0	0.82547226862385
+0	1 24 24	0	-7.6666666666667 0 0	0.81857245720431
+0	2 24 24	0	-7.3333333333333 0 0	0.81110463303799
+0	3 24 24	0	-7 0 0	0.8029956443743
+0	4 24 24	0	-6.6666666666667 0 0	0.7941592148739
+0	5 24 24	0	-6.3333333333333 0 0	0.78449286193305
+0	6 24 24	0	-6 0 0	0.77387392843168
+0	7 24 24	0	-5.6666666666667 0 0	0.76218721796563
+0	8 24 24	0	-5.3333333333333 0 0	0.74964251213876
+0	9 24 24	0	-5 0 0	0.73653458961075
+0	10 24 24	0	-4.6666666666667 0 0	0.72305160410817
+0	11 24 24	0	-4.3333333333333 0 0	0.70936213930298
+0	12 24 24	0	-4 0 0	0.69563187399132
+0	13 24 24	0	-3.6666666666667 0 0	0.68203032240126
+0	14 24 24	0	-3.3333333333333 0 0	0.66873262281794
+0	15 24 24	0	-3 0 0	0.65591962514845
+0	16 24 24	0	-2.6666666666667 0 0	0.64377655892996
+0	17 24 24	0	-2.3333333333333 0 0	0.63249037820859
+0	18 24 24	0	-2 0 0	0.62224591766624
+0	19 24 24	0	-1.6666666666667 0 0	0.61322102249476
+0	20 24 24	0	-1.3333333333333 0 0	0.60558089769349
+0	21 24 24	0	-1 0 0	0.59947202998127
+0	22 24 24	0	-0.66666666666667 0 0	0.59501613860795
+0	23 24 24	0	-0.33333333333333 0 0	0.59230468026646
+0	24 24 24	0	0 0 0	0.59139450673206
+0	25 24 24	0	0.33333333333333 0 0	0.59230468026646
+0	26 24 24	0	0.66666666666667 0 0	0.59501613860795
+0	27 24 24	0	1 0 0	0.59947202998127
+0	28 24 24	0	1.3333333333333 0 0	0.60558089769349
+0	29 24 24	0	1.6666666666667 0 0	0.61322102249476
+0	30 24 24	0	2 0 0	0.62224591766624
+0	31 24 24	0	2.3333333333333 0 0	0.63249037820859
+0	32 24 24	0	2.6666666666667 0 0	0.64377655892996
+0	33 24 24	0	3 0 0	0.65591962514845
+0	34 24 24	0	3.3333333333333 0 0	0.66873262281794
+0	35 24 24	0	3.6666666666667 0 0	0.68203032240126
+0	36 24 24	0	4 0 0	0.69563187399132
+0	37 24 24	0	4.3333333333333 0 0	0.70936213930298
+0	38 24 24	0	4.6666666666667 0 0	0.72305160410817
+0	39 24 24	0	5 0 0	0.73653458961075
+0	40 24 24	0	5.3333333333333 0 0	0.74964251213876
+0	41 24 24	0	5.6666666666667 0 0	0.76218721800381
+0	42 24 24	0	6 0 0	0.77387392843168
+0	43 24 24	0	6.3333333333333 0 0	0.78449286193305
+0	44 24 24	0	6.6666666666667 0 0	0.7941592148739
+0	45 24 24	0	7 0 0	0.8029956443743
+0	46 24 24	0	7.3333333333333 0 0	0.81110463303799
+0	47 24 24	0	7.6666666666667 0 0	0.81857245720431
+0	48 24 24	0	8 0 0	0.82547226862385
