+++ TOVola/test/Piecewise_test/eps.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-8 0 0	7.3273025657041e-06
-0	1 24 24	0	-7.6666666666667 0 0	7.3273025657041e-06
-0	2 24 24	0	-7.3333333333333 0 0	7.3273025660561e-06
-0	3 24 24	0	-7 0 0	7.3273025660561e-06
-0	4 24 24	0	-6.6666666666667 0 0	7.3273025657041e-06
-0	5 24 24	0	-6.3333333333333 0 0	7.3273025657041e-06
-0	6 24 24	0	-6 0 0	0.0078773103485269
-0	7 24 24	0	-5.6666666666667 0 0	0.020582119625703
-0	8 24 24	0	-5.3333333333333 0 0	0.026292973233336
-0	9 24 24	0	-5 0 0	0.032468062095615
-0	10 24 24	0	-4.6666666666667 0 0	0.039053434304871
-0	11 24 24	0	-4.3333333333333 0 0	0.046001829137433
-0	12 24 24	0	-4 0 0	0.05328049921777
-0	13 24 24	0	-3.6666666666667 0 0	0.060779888199436
-0	14 24 24	0	-3.3333333333333 0 0	0.068406662735512
-0	15 24 24	0	-3 0 0	0.07604797936931
-0	16 24 24	0	-2.6666666666667 0 0	0.083570496109485
-0	17 24 24	0	-2.3333333333333 0 0	0.090821172157786
-0	18 24 24	0	-2 0 0	0.097630334659904
-0	19 24 24	0	-1.6666666666667 0 0	0.10381738513479
-0	20 24 24	0	-1.3333333333333 0 0	0.10919923817445
-0	21 24 24	0	-1 0 0	0.11360113858207
-0	22 24 24	0	-0.66666666666667 0 0	0.11686895393655
-0	23 24 24	0	-0.33333333333333 0 0	0.11888151752363
-0	24 24 24	0	0 0 0	0.11956127443508
-0	25 24 24	0	0.33333333333333 0 0	0.11888151752363
-0	26 24 24	0	0.66666666666667 0 0	0.11686895393655
-0	27 24 24	0	1 0 0	0.11360113858207
-0	28 24 24	0	1.3333333333333 0 0	0.10919923817445
-0	29 24 24	0	1.6666666666667 0 0	0.1038173851348
-0	30 24 24	0	2 0 0	0.097630334659904
-0	31 24 24	0	2.3333333333333 0 0	0.090821172157786
-0	32 24 24	0	2.6666666666667 0 0	0.083570496109485
-0	33 24 24	0	3 0 0	0.07604797936931
-0	34 24 24	0	3.3333333333333 0 0	0.068406662735511
-0	35 24 24	0	3.6666666666667 0 0	0.060779888199436
-0	36 24 24	0	4 0 0	0.05328049921777
-0	37 24 24	0	4.3333333333333 0 0	0.046001829137433
-0	38 24 24	0	4.6666666666667 0 0	0.039053434304871
-0	39 24 24	0	5 0 0	0.032468062095615
-0	40 24 24	0	5.3333333333333 0 0	0.026292973233336
-0	41 24 24	0	5.6666666666667 0 0	0.020582119625704
-0	42 24 24	0	6 0 0	0.0078773103485269
-0	43 24 24	0	6.3333333333333 0 0	7.3273025657041e-06
+0	0 24 24	0	-8 0 0	7.3273025657406e-06
+0	1 24 24	0	-7.6666666666667 0 0	7.3273025663086e-06
+0	2 24 24	0	-7.3333333333333 0 0	7.3273025661823e-06
+0	3 24 24	0	-7 0 0	7.3273025657041e-06
+0	4 24 24	0	-6.6666666666667 0 0	7.3273025659299e-06
+0	5 24 24	0	-6.3333333333333 0 0	7.3273025664348e-06
+0	6 24 24	0	-6 0 0	0.0078773016949452
+0	7 24 24	0	-5.6666666666667 0 0	0.020600513966528
+0	8 24 24	0	-5.3333333333333 0 0	0.026292973175877
+0	9 24 24	0	-5 0 0	0.032468062028126
+0	10 24 24	0	-4.6666666666667 0 0	0.039053435378703
+0	11 24 24	0	-4.3333333333333 0 0	0.046001855840023
+0	12 24 24	0	-4 0 0	0.053280499244154
+0	13 24 24	0	-3.6666666666667 0 0	0.0607798882242
+0	14 24 24	0	-3.3333333333333 0 0	0.068406662758245
+0	15 24 24	0	-3 0 0	0.07604797938964
+0	16 24 24	0	-2.6666666666667 0 0	0.083570496127099
+0	17 24 24	0	-2.3333333333333 0 0	0.090821172172462
+0	18 24 24	0	-2 0 0	0.09763033467154
+0	19 24 24	0	-1.6666666666667 0 0	0.10381738514343
+0	20 24 24	0	-1.3333333333333 0 0	0.10919923818029
+0	21 24 24	0	-1 0 0	0.1136011385855
+0	22 24 24	0	-0.66666666666667 0 0	0.11686895393813
+0	23 24 24	0	-0.33333333333333 0 0	0.11888151752403
+0	24 24 24	0	0 0 0	0.11956131307275
+0	25 24 24	0	0.33333333333333 0 0	0.11888151752403
+0	26 24 24	0	0.66666666666667 0 0	0.11686895393813
+0	27 24 24	0	1 0 0	0.1136011385855
+0	28 24 24	0	1.3333333333333 0 0	0.10919923818029
+0	29 24 24	0	1.6666666666667 0 0	0.10381738514343
+0	30 24 24	0	2 0 0	0.09763033467154
+0	31 24 24	0	2.3333333333333 0 0	0.090821172172462
+0	32 24 24	0	2.6666666666667 0 0	0.0835704961271
+0	33 24 24	0	3 0 0	0.07604797938964
+0	34 24 24	0	3.3333333333333 0 0	0.068406662758246
+0	35 24 24	0	3.6666666666667 0 0	0.0607798882242
+0	36 24 24	0	4 0 0	0.053280499244154
+0	37 24 24	0	4.3333333333333 0 0	0.046001855840023
+0	38 24 24	0	4.6666666666667 0 0	0.039053435378704
+0	39 24 24	0	5 0 0	0.032468062028126
+0	40 24 24	0	5.3333333333333 0 0	0.026292973175877
+0	41 24 24	0	5.6666666666667 0 0	0.020600513966985
+0	42 24 24	0	6 0 0	0.0078773016949452
+0	43 24 24	0	6.3333333333333 0 0	7.3273025658037e-06
 0	44 24 24	0	6.6666666666667 0 0	7.3273025657041e-06
-0	45 24 24	0	7 0 0	7.3273025660561e-06
-0	46 24 24	0	7.3333333333333 0 0	7.3273025660561e-06
-0	47 24 24	0	7.6666666666667 0 0	7.3273025657041e-06
-0	48 24 24	0	8 0 0	7.3273025657041e-06
+0	45 24 24	0	7 0 0	7.3273025657041e-06
+0	46 24 24	0	7.3333333333333 0 0	7.3273025658037e-06
+0	47 24 24	0	7.6666666666667 0 0	7.3273025663086e-06
+0	48 24 24	0	8 0 0	7.3273025657406e-06
