+++ TOVola/test/Piecewise_test/gyy.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-8 0 0	1.4408508385388
-0	1 24 24	0	-7.6666666666667 0 0	1.4628424202074
-0	2 24 24	0	-7.3333333333333 0 0	1.4871193228874
-0	3 24 24	0	-7 0 0	1.5140536782715
-0	4 24 24	0	-6.6666666666667 0 0	1.5441022752985
-0	5 24 24	0	-6.3333333333333 0 0	1.5778318648842
-0	6 24 24	0	-6 0 0	1.6159539361761
-0	7 24 24	0	-5.6666666666667 0 0	1.6592322603558
-0	8 24 24	0	-5.3333333333333 0 0	1.7070925551241
-0	9 24 24	0	-5 0 0	1.7585313934123
-0	10 24 24	0	-4.6666666666667 0 0	1.8129428081559
-0	11 24 24	0	-4.3333333333333 0 0	1.8697517559085
-0	12 24 24	0	-4 0 0	1.9283362912108
-0	13 24 24	0	-3.6666666666667 0 0	1.9879889761784
-0	14 24 24	0	-3.3333333333333 0 0	2.047900695483
-0	15 24 24	0	-3 0 0	2.1071537157618
-0	16 24 24	0	-2.6666666666667 0 0	2.1647247394407
-0	17 24 24	0	-2.3333333333333 0 0	2.2194995088238
-0	18 24 24	0	-2 0 0	2.2702999956162
-0	19 24 24	0	-1.6666666666667 0 0	2.3159244547355
-0	20 24 24	0	-1.3333333333333 0 0	2.355199202992
-0	21 24 24	0	-1 0 0	2.3870391022141
-0	22 24 24	0	-0.66666666666667 0 0	2.410511727209
-0	23 24 24	0	-0.33333333333333 0 0	2.4248985705386
-0	24 24 24	0	0 0 0	2.4297459125258
-0	25 24 24	0	0.33333333333333 0 0	2.4248985705386
-0	26 24 24	0	0.66666666666667 0 0	2.410511727209
-0	27 24 24	0	1 0 0	2.3870391022141
-0	28 24 24	0	1.3333333333333 0 0	2.355199202992
-0	29 24 24	0	1.6666666666667 0 0	2.3159244547355
-0	30 24 24	0	2 0 0	2.2702999956162
-0	31 24 24	0	2.3333333333333 0 0	2.2194995088238
-0	32 24 24	0	2.6666666666667 0 0	2.1647247394407
-0	33 24 24	0	3 0 0	2.1071537157618
-0	34 24 24	0	3.3333333333333 0 0	2.047900695483
-0	35 24 24	0	3.6666666666667 0 0	1.9879889761784
-0	36 24 24	0	4 0 0	1.9283362912108
-0	37 24 24	0	4.3333333333333 0 0	1.8697517559085
-0	38 24 24	0	4.6666666666667 0 0	1.8129428081559
-0	39 24 24	0	5 0 0	1.7585313934123
-0	40 24 24	0	5.3333333333333 0 0	1.7070925551241
-0	41 24 24	0	5.6666666666667 0 0	1.6592322603558
-0	42 24 24	0	6 0 0	1.6159539361761
-0	43 24 24	0	6.3333333333333 0 0	1.5778318648842
-0	44 24 24	0	6.6666666666667 0 0	1.5441022752985
-0	45 24 24	0	7 0 0	1.5140536782715
-0	46 24 24	0	7.3333333333333 0 0	1.4871193228874
-0	47 24 24	0	7.6666666666667 0 0	1.4628424202074
-0	48 24 24	0	8 0 0	1.4408508385388
+0	0 24 24	0	-8 0 0	1.440850837822
+0	1 24 24	0	-7.6666666666667 0 0	1.4628424194509
+0	2 24 24	0	-7.3333333333333 0 0	1.4871193220867
+0	3 24 24	0	-7 0 0	1.5140536774212
+0	4 24 24	0	-6.6666666666667 0 0	1.5441022743925
+0	5 24 24	0	-6.3333333333333 0 0	1.5778318639149
+0	6 24 24	0	-6 0 0	1.6159539351348
+0	7 24 24	0	-5.6666666666667 0 0	1.6592322235228
+0	8 24 24	0	-5.3333333333333 0 0	1.7070925541724
+0	9 24 24	0	-5 0 0	1.7585313924644
+0	10 24 24	0	-4.6666666666667 0 0	1.8129428072338
+0	11 24 24	0	-4.3333333333333 0 0	1.8697517550114
+0	12 24 24	0	-4 0 0	1.9283362902766
+0	13 24 24	0	-3.6666666666667 0 0	1.9879889751934
+0	14 24 24	0	-3.3333333333333 0 0	2.047900694444
+0	15 24 24	0	-3 0 0	2.1071537146668
+0	16 24 24	0	-2.6666666666667 0 0	2.1647247382888
+0	17 24 24	0	-2.3333333333333 0 0	2.2194995076157
+0	18 24 24	0	-2 0 0	2.270299994354
+0	19 24 24	0	-1.6666666666667 0 0	2.3159244534233
+0	20 24 24	0	-1.3333333333333 0 0	2.3551992016356
+0	21 24 24	0	-1 0 0	2.3870391008211
+0	22 24 24	0	-0.66666666666667 0 0	2.4105117257886
+0	23 24 24	0	-0.33333333333333 0 0	2.4248985691012
+0	24 24 24	0	0 0 0	2.429746330643
+0	25 24 24	0	0.33333333333333 0 0	2.4248985691012
+0	26 24 24	0	0.66666666666667 0 0	2.4105117257886
+0	27 24 24	0	1 0 0	2.3870391008211
+0	28 24 24	0	1.3333333333333 0 0	2.3551992016356
+0	29 24 24	0	1.6666666666667 0 0	2.3159244534233
+0	30 24 24	0	2 0 0	2.270299994354
+0	31 24 24	0	2.3333333333333 0 0	2.2194995076157
+0	32 24 24	0	2.6666666666667 0 0	2.1647247382888
+0	33 24 24	0	3 0 0	2.1071537146668
+0	34 24 24	0	3.3333333333333 0 0	2.047900694444
+0	35 24 24	0	3.6666666666667 0 0	1.9879889751934
+0	36 24 24	0	4 0 0	1.9283362902766
+0	37 24 24	0	4.3333333333333 0 0	1.8697517550114
+0	38 24 24	0	4.6666666666667 0 0	1.8129428072338
+0	39 24 24	0	5 0 0	1.7585313924644
+0	40 24 24	0	5.3333333333333 0 0	1.7070925541724
+0	41 24 24	0	5.6666666666667 0 0	1.65923222329
+0	42 24 24	0	6 0 0	1.6159539351348
+0	43 24 24	0	6.3333333333333 0 0	1.5778318639149
+0	44 24 24	0	6.6666666666667 0 0	1.5441022743925
+0	45 24 24	0	7 0 0	1.5140536774212
+0	46 24 24	0	7.3333333333333 0 0	1.4871193220867
+0	47 24 24	0	7.6666666666667 0 0	1.4628424194509
+0	48 24 24	0	8 0 0	1.440850837822
