+++ TOVola/test/Simple_test/HGF.x.asc
@@ -5,45 +5,45 @@
 0	2 24 24	0	-1.8333333333333 0 0	-nan
 0	3 24 24	0	-1.75 0 0	-nan
 0	4 24 24	0	-1.6666666666667 0 0	-nan
-0	5 24 24	0	-1.5833333333333 0 0	1.7776233427602e-06
-0	6 24 24	0	-1.5 0 0	2.5034314521259e-06
-0	7 24 24	0	-1.4166666666667 0 0	3.5871836229079e-06
-0	8 24 24	0	-1.3333333333333 0 0	5.239256605739e-06
-0	9 24 24	0	-1.25 0 0	7.8159007864263e-06
-0	10 24 24	0	-1.1666666666667 0 0	1.1937577815163e-05
-0	11 24 24	0	-1.0833333333333 0 0	1.871880305045e-05
-0	12 24 24	0	-1 0 0	3.023040961648e-05
-0	13 24 24	0	-0.91666666666667 0 0	-0.00029092837312518
-0	14 24 24	0	-0.83333333333333 0 0	-0.01623996356851
-0	15 24 24	0	-0.75 0 0	0.0026219292213432
-0	16 24 24	0	-0.66666666666667 0 0	-0.0020187185053488
-0	17 24 24	0	-0.58333333333333 0 0	-0.00046513374521151
-0	18 24 24	0	-0.5 0 0	-0.00049081802755246
-0	19 24 24	0	-0.41666666666667 0 0	-0.00047606795545097
-0	20 24 24	0	-0.33333333333333 0 0	-0.00052393944268569
-0	21 24 24	0	-0.25 0 0	-0.00091956142358729
-0	22 24 24	0	-0.16666666666667 0 0	-0.0019552041880928
-0	23 24 24	0	-0.083333333333333 0 0	-0.003345631315284
-0	24 24 24	0	0 0 0	-0.0040341499899501
-0	25 24 24	0	0.083333333333333 0 0	-0.0033456313141418
-0	26 24 24	0	0.16666666666667 0 0	-0.0019552041886506
-0	27 24 24	0	0.25 0 0	-0.00091956142356331
-0	28 24 24	0	0.33333333333333 0 0	-0.00052393944181883
-0	29 24 24	0	0.41666666666667 0 0	-0.00047606795611221
-0	30 24 24	0	0.5 0 0	-0.00049081802719675
-0	31 24 24	0	0.58333333333333 0 0	-0.00046513374556234
-0	32 24 24	0	0.66666666666667 0 0	-0.0020187185041435
-0	33 24 24	0	0.75 0 0	0.0026219292211223
-0	34 24 24	0	0.83333333333333 0 0	-0.016239963568513
-0	35 24 24	0	0.91666666666667 0 0	-0.00029092837312387
-0	36 24 24	0	1 0 0	3.0230409613691e-05
-0	37 24 24	0	1.0833333333333 0 0	1.8718802328924e-05
-0	38 24 24	0	1.1666666666667 0 0	1.1937579490416e-05
-0	39 24 24	0	1.25 0 0	7.8159007320786e-06
-0	40 24 24	0	1.3333333333333 0 0	5.239256603379e-06
-0	41 24 24	0	1.4166666666667 0 0	3.5871836243606e-06
-0	42 24 24	0	1.5 0 0	2.5034314520341e-06
-0	43 24 24	0	1.5833333333333 0 0	1.7776233190306e-06
+0	5 24 24	0	-1.5833333333333 0 0	1.7558320068167e-06
+0	6 24 24	0	-1.5 0 0	2.4792743932902e-06
+0	7 24 24	0	-1.4166666666667 0 0	3.5633335282067e-06
+0	8 24 24	0	-1.3333333333333 0 0	5.222705144877e-06
+0	9 24 24	0	-1.25 0 0	7.8233139523126e-06
+0	10 24 24	0	-1.1666666666667 0 0	1.2007485797489e-05
+0	11 24 24	0	-1.0833333333333 0 0	1.8940120681637e-05
+0	12 24 24	0	-1 0 0	3.0811961430034e-05
+0	13 24 24	0	-0.91666666666667 0 0	-0.0038334917158657
+0	14 24 24	0	-0.83333333333333 0 0	0.013401396410258
+0	15 24 24	0	-0.75 0 0	-0.00095230825958587
+0	16 24 24	0	-0.66666666666667 0 0	-0.00028572904121971
+0	17 24 24	0	-0.58333333333333 0 0	-0.00027442647876041
+0	18 24 24	0	-0.5 0 0	-0.00027410670034689
+0	19 24 24	0	-0.41666666666667 0 0	-0.00027670774423427
+0	20 24 24	0	-0.33333333333333 0 0	-0.00036438075695155
+0	21 24 24	0	-0.25 0 0	-0.00067484271088336
+0	22 24 24	0	-0.16666666666667 0 0	-0.0012820559062448
+0	23 24 24	0	-0.083333333333333 0 0	-0.0018893590245117
+0	24 24 24	0	0 0 0	-0.0026840950106921
+0	25 24 24	0	0.083333333333333 0 0	-0.0018893590247009
+0	26 24 24	0	0.16666666666667 0 0	-0.0012820559056221
+0	27 24 24	0	0.25 0 0	-0.00067484271086826
+0	28 24 24	0	0.33333333333333 0 0	-0.0003643807571807
+0	29 24 24	0	0.41666666666667 0 0	-0.00027670774300104
+0	30 24 24	0	0.5 0 0	-0.00027410670075367
+0	31 24 24	0	0.58333333333333 0 0	-0.0002744264785568
+0	32 24 24	0	0.66666666666667 0 0	-0.00028572904095514
+0	33 24 24	0	0.75 0 0	-0.00095230825963055
+0	34 24 24	0	0.83333333333333 0 0	0.013401396410266
+0	35 24 24	0	0.91666666666667 0 0	-0.0038334917158644
+0	36 24 24	0	1 0 0	3.0811961429949e-05
+0	37 24 24	0	1.0833333333333 0 0	1.8940120681637e-05
+0	38 24 24	0	1.1666666666667 0 0	1.200748579467e-05
+0	39 24 24	0	1.25 0 0	7.8233139523126e-06
+0	40 24 24	0	1.3333333333333 0 0	5.2227051447864e-06
+0	41 24 24	0	1.4166666666667 0 0	3.5633335281151e-06
+0	42 24 24	0	1.5 0 0	2.4792743932902e-06
+0	43 24 24	0	1.5833333333333 0 0	1.75583200677e-06
 0	44 24 24	0	1.6666666666667 0 0	-nan
 0	45 24 24	0	1.75 0 0	-nan
 0	46 24 24	0	1.8333333333333 0 0	-nan
