+++ TOVola/test/Simple_test/alp.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-2 0 0	0.92951818180567
-0	1 24 24	0	-1.9166666666667 0 0	0.92656627107399
-0	2 24 24	0	-1.8333333333333 0 0	0.92335628870279
-0	3 24 24	0	-1.75 0 0	0.9198528447227
-0	4 24 24	0	-1.6666666666667 0 0	0.91601376839747
-0	5 24 24	0	-1.5833333333333 0 0	0.91178840251461
-0	6 24 24	0	-1.5 0 0	0.90711535551187
-0	7 24 24	0	-1.4166666666667 0 0	0.90191949914355
-0	8 24 24	0	-1.3333333333333 0 0	0.89610789875642
-0	9 24 24	0	-1.25 0 0	0.88956420556473
-0	10 24 24	0	-1.1666666666667 0 0	0.88214078724052
-0	11 24 24	0	-1.0833333333333 0 0	0.87364745592727
-0	12 24 24	0	-1 0 0	0.863834943973
-0	13 24 24	0	-0.91666666666667 0 0	0.8523700317387
-0	14 24 24	0	-0.83333333333333 0 0	0.83879695493391
-0	15 24 24	0	-0.75 0 0	0.82248332006564
-0	16 24 24	0	-0.66666666666667 0 0	0.80299436706974
-0	17 24 24	0	-0.58333333333333 0 0	0.78055822209856
-0	18 24 24	0	-0.5 0 0	0.75573802430975
-0	19 24 24	0	-0.41666666666667 0 0	0.72949156162671
-0	20 24 24	0	-0.33333333333333 0 0	0.70326518231358
-0	21 24 24	0	-0.25 0 0	0.67902261007764
-0	22 24 24	0	-0.16666666666667 0 0	0.65911966856159
-0	23 24 24	0	-0.083333333333333 0 0	0.64594069023674
-0	24 24 24	0	0 0 0	0.64131642215947
-0	25 24 24	0	0.083333333333333 0 0	0.64594069023674
-0	26 24 24	0	0.16666666666667 0 0	0.65911966856159
-0	27 24 24	0	0.25 0 0	0.67902261007764
-0	28 24 24	0	0.33333333333333 0 0	0.70326518231358
-0	29 24 24	0	0.41666666666667 0 0	0.72949156162671
-0	30 24 24	0	0.5 0 0	0.75573802430975
-0	31 24 24	0	0.58333333333333 0 0	0.78055822209856
-0	32 24 24	0	0.66666666666667 0 0	0.80299436706974
-0	33 24 24	0	0.75 0 0	0.82248332006564
-0	34 24 24	0	0.83333333333333 0 0	0.83879695493391
-0	35 24 24	0	0.91666666666667 0 0	0.8523700317387
-0	36 24 24	0	1 0 0	0.863834943973
-0	37 24 24	0	1.0833333333333 0 0	0.87364745592727
-0	38 24 24	0	1.1666666666667 0 0	0.88214078724052
-0	39 24 24	0	1.25 0 0	0.88956420556473
-0	40 24 24	0	1.3333333333333 0 0	0.89610789875642
-0	41 24 24	0	1.4166666666667 0 0	0.90191949914355
-0	42 24 24	0	1.5 0 0	0.90711535551187
-0	43 24 24	0	1.5833333333333 0 0	0.91178840251461
-0	44 24 24	0	1.6666666666667 0 0	0.91601376839747
-0	45 24 24	0	1.75 0 0	0.9198528447227
-0	46 24 24	0	1.8333333333333 0 0	0.92335628870279
-0	47 24 24	0	1.9166666666667 0 0	0.92656627107399
-0	48 24 24	0	2 0 0	0.92951818180567
+0	0 24 24	0	-2 0 0	0.93290647533218
+0	1 24 24	0	-1.9166666666667 0 0	0.93009133119435
+0	2 24 24	0	-1.8333333333333 0 0	0.92702960223432
+0	3 24 24	0	-1.75 0 0	0.92368740615803
+0	4 24 24	0	-1.6666666666667 0 0	0.92002435518842
+0	5 24 24	0	-1.5833333333333 0 0	0.91599191600589
+0	6 24 24	0	-1.5 0 0	0.91153124718447
+0	7 24 24	0	-1.4166666666667 0 0	0.90657030902581
+0	8 24 24	0	-1.3333333333333 0 0	0.9010199432186
+0	9 24 24	0	-1.25 0 0	0.89476846687065
+0	10 24 24	0	-1.1666666666667 0 0	0.88767407981823
+0	11 24 24	0	-1.0833333333333 0 0	0.87955397865731
+0	12 24 24	0	-1 0 0	0.87016838112616
+0	13 24 24	0	-0.91666666666667 0 0	0.85919644999401
+0	14 24 24	0	-0.83333333333333 0 0	0.84619888216273
+0	15 24 24	0	-0.75 0 0	0.83064848171515
+0	16 24 24	0	-0.66666666666667 0 0	0.81250423367303
+0	17 24 24	0	-0.58333333333333 0 0	0.79212190438128
+0	18 24 24	0	-0.5 0 0	0.77010887016778
+0	19 24 24	0	-0.41666666666667 0 0	0.74738328702039
+0	20 24 24	0	-0.33333333333333 0 0	0.7252105987147
+0	21 24 24	0	-0.25 0 0	0.70517821021673
+0	22 24 24	0	-0.16666666666667 0 0	0.68906135831592
+0	23 24 24	0	-0.083333333333333 0 0	0.67855299038502
+0	24 24 24	0	0 0 0	0.67489658522026
+0	25 24 24	0	0.083333333333333 0 0	0.67855299038502
+0	26 24 24	0	0.16666666666667 0 0	0.68906135831592
+0	27 24 24	0	0.25 0 0	0.70517821021673
+0	28 24 24	0	0.33333333333333 0 0	0.7252105987147
+0	29 24 24	0	0.41666666666667 0 0	0.74738328702039
+0	30 24 24	0	0.5 0 0	0.77010887016778
+0	31 24 24	0	0.58333333333333 0 0	0.79212190438128
+0	32 24 24	0	0.66666666666667 0 0	0.81250423367303
+0	33 24 24	0	0.75 0 0	0.83064848171515
+0	34 24 24	0	0.83333333333333 0 0	0.84619888216273
+0	35 24 24	0	0.91666666666667 0 0	0.85919644999401
+0	36 24 24	0	1 0 0	0.87016838112616
+0	37 24 24	0	1.0833333333333 0 0	0.87955397865731
+0	38 24 24	0	1.1666666666667 0 0	0.88767407981823
+0	39 24 24	0	1.25 0 0	0.89476846687065
+0	40 24 24	0	1.3333333333333 0 0	0.9010199432186
+0	41 24 24	0	1.4166666666667 0 0	0.90657030902581
+0	42 24 24	0	1.5 0 0	0.91153124718447
+0	43 24 24	0	1.5833333333333 0 0	0.91599191600589
+0	44 24 24	0	1.6666666666667 0 0	0.92002435518842
+0	45 24 24	0	1.75 0 0	0.92368740615803
+0	46 24 24	0	1.8333333333333 0 0	0.92702960223432
+0	47 24 24	0	1.9166666666667 0 0	0.93009133119435
+0	48 24 24	0	2 0 0	0.93290647533218
