+++ TOVola/test/Simple_test/eps.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-2 0 0	1.0010844469964e-13
+0	0 24 24	0	-2 0 0	1e-13
 0	1 24 24	0	-1.9166666666667 0 0	1e-13
 0	2 24 24	0	-1.8333333333333 0 0	1e-13
 0	3 24 24	0	-1.75 0 0	1e-13
 0	4 24 24	0	-1.6666666666667 0 0	1e-13
 0	5 24 24	0	-1.5833333333333 0 0	1e-13
-0	6 24 24	0	-1.5 0 0	1e-13
-0	7 24 24	0	-1.4166666666667 0 0	1e-13
+0	6 24 24	0	-1.5 0 0	1.0023466244447e-13
+0	7 24 24	0	-1.4166666666667 0 0	1.0010844469964e-13
 0	8 24 24	0	-1.3333333333333 0 0	1e-13
 0	9 24 24	0	-1.25 0 0	1e-13
-0	10 24 24	0	-1.1666666666667 0 0	1e-13
-0	11 24 24	0	-1.0833333333333 0 0	1e-13
+0	10 24 24	0	-1.1666666666667 0 0	1.0023466244447e-13
+0	11 24 24	0	-1.0833333333333 0 0	1.0010844469964e-13
 0	12 24 24	0	-1 0 0	1e-13
-0	13 24 24	0	-0.91666666666667 0 0	1.0029777131689e-13
-0	14 24 24	0	-0.83333333333333 0 0	1e-13
-0	15 24 24	0	-0.75 0 0	0.0037064573591128
-0	16 24 24	0	-0.66666666666667 0 0	0.015931588535343
-0	17 24 24	0	-0.58333333333333 0 0	0.030761380327775
-0	18 24 24	0	-0.5 0 0	0.048192820872831
-0	19 24 24	0	-0.41666666666667 0 0	0.067916314841793
-0	20 24 24	0	-0.33333333333333 0 0	0.089095222977351
-0	21 24 24	0	-0.25 0 0	0.11012719346688
-0	22 24 24	0	-0.16666666666667 0 0	0.12855074601451
-0	23 24 24	0	-0.083333333333333 0 0	0.14137492133432
-0	24 24 24	0	0 0 0	0.145999611225
-0	25 24 24	0	0.083333333333333 0 0	0.14137492133432
-0	26 24 24	0	0.16666666666667 0 0	0.12855074601451
-0	27 24 24	0	0.25 0 0	0.11012719346688
-0	28 24 24	0	0.33333333333333 0 0	0.089095222977351
-0	29 24 24	0	0.41666666666667 0 0	0.067916314841794
-0	30 24 24	0	0.5 0 0	0.048192820872831
-0	31 24 24	0	0.58333333333333 0 0	0.030761380327774
-0	32 24 24	0	0.66666666666667 0 0	0.015931588535343
-0	33 24 24	0	0.75 0 0	0.0037064573591128
+0	13 24 24	0	-0.91666666666667 0 0	1e-13
+0	14 24 24	0	-0.83333333333333 0 0	1.0036088018931e-13
+0	15 24 24	0	-0.75 0 0	0.0078086560876132
+0	16 24 24	0	-0.66666666666667 0 0	0.019148666197691
+0	17 24 24	0	-0.58333333333333 0 0	0.03250703819484
+0	18 24 24	0	-0.5 0 0	0.047728386896039
+0	19 24 24	0	-0.41666666666667 0 0	0.064383090332369
+0	20 24 24	0	-0.33333333333333 0 0	0.081638616339732
+0	21 24 24	0	-0.25 0 0	0.098161547081395
+0	22 24 24	0	-0.16666666666667 0 0	0.11215229108514
+0	23 24 24	0	-0.083333333333333 0 0	0.12163234878973
+0	24 24 24	0	0 0 0	0.125
+0	25 24 24	0	0.083333333333333 0 0	0.12163234878972
+0	26 24 24	0	0.16666666666667 0 0	0.11215229108514
+0	27 24 24	0	0.25 0 0	0.098161547081395
+0	28 24 24	0	0.33333333333333 0 0	0.081638616339733
+0	29 24 24	0	0.41666666666667 0 0	0.064383090332369
+0	30 24 24	0	0.5 0 0	0.047728386896039
+0	31 24 24	0	0.58333333333333 0 0	0.032507038194839
+0	32 24 24	0	0.66666666666667 0 0	0.01914866619769
+0	33 24 24	0	0.75 0 0	0.0078086560876132
 0	34 24 24	0	0.83333333333333 0 0	1e-13
 0	35 24 24	0	0.91666666666667 0 0	1e-13
 0	36 24 24	0	1 0 0	1e-13
-0	37 24 24	0	1.0833333333333 0 0	1e-13
-0	38 24 24	0	1.1666666666667 0 0	1e-13
+0	37 24 24	0	1.0833333333333 0 0	1.0010844469964e-13
+0	38 24 24	0	1.1666666666667 0 0	1.0023466244447e-13
 0	39 24 24	0	1.25 0 0	1e-13
 0	40 24 24	0	1.3333333333333 0 0	1e-13
-0	41 24 24	0	1.4166666666667 0 0	1e-13
-0	42 24 24	0	1.5 0 0	1e-13
+0	41 24 24	0	1.4166666666667 0 0	1.0010844469964e-13
+0	42 24 24	0	1.5 0 0	1.0023466244447e-13
 0	43 24 24	0	1.5833333333333 0 0	1e-13
 0	44 24 24	0	1.6666666666667 0 0	1e-13
 0	45 24 24	0	1.75 0 0	1e-13
 0	46 24 24	0	1.8333333333333 0 0	1e-13
 0	47 24 24	0	1.9166666666667 0 0	1e-13
-0	48 24 24	0	2 0 0	1.0010844469964e-13
+0	48 24 24	0	2 0 0	1e-13
