+++ TOVola/test/Simple_test/gxx.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-2 0 0	1.1543153783021
-0	1 24 24	0	-1.9166666666667 0 0	1.1614062848245
-0	2 24 24	0	-1.8333333333333 0 0	1.1691790248092
-0	3 24 24	0	-1.75 0 0	1.1777367206214
-0	4 24 24	0	-1.6666666666667 0 0	1.1872043530817
-0	5 24 24	0	-1.5833333333333 0 0	1.1977348653292
-0	6 24 24	0	-1.5 0 0	1.209517428715
-0	7 24 24	0	-1.4166666666667 0 0	1.2227888182481
-0	8 24 24	0	-1.3333333333333 0 0	1.2378493502812
-0	9 24 24	0	-1.25 0 0	1.2550856641843
-0	10 24 24	0	-1.1666666666667 0 0	1.2750040305068
-0	11 24 24	0	-1.0833333333333 0 0	1.2982803136804
-0	12 24 24	0	-1 0 0	1.325837148809
-0	13 24 24	0	-0.91666666666667 0 0	1.3589672721216
-0	14 24 24	0	-0.83333333333333 0 0	1.3995385938454
-0	15 24 24	0	-0.75 0 0	1.4503238108871
-0	16 24 24	0	-0.66666666666667 0 0	1.5139033453782
-0	17 24 24	0	-0.58333333333333 0 0	1.5909540344048
-0	18 24 24	0	-0.5 0 0	1.6811041609329
-0	19 24 24	0	-0.41666666666667 0 0	1.7823122145957
-0	20 24 24	0	-0.33333333333333 0 0	1.889849516812
-0	21 24 24	0	-0.25 0 0	1.9953343559667
-0	22 24 24	0	-0.16666666666667 0 0	2.0865912899806
-0	23 24 24	0	-0.083333333333333 0 0	2.1494558010727
-0	24 24 24	0	0 0 0	2.1719911372882
-0	25 24 24	0	0.083333333333333 0 0	2.1494558010727
-0	26 24 24	0	0.16666666666667 0 0	2.0865912899806
-0	27 24 24	0	0.25 0 0	1.9953343559667
-0	28 24 24	0	0.33333333333333 0 0	1.889849516812
-0	29 24 24	0	0.41666666666667 0 0	1.7823122145957
-0	30 24 24	0	0.5 0 0	1.6811041609329
-0	31 24 24	0	0.58333333333333 0 0	1.5909540344048
-0	32 24 24	0	0.66666666666667 0 0	1.5139033453782
-0	33 24 24	0	0.75 0 0	1.4503238108871
-0	34 24 24	0	0.83333333333333 0 0	1.3995385938454
-0	35 24 24	0	0.91666666666667 0 0	1.3589672721216
-0	36 24 24	0	1 0 0	1.325837148809
-0	37 24 24	0	1.0833333333333 0 0	1.2982803136804
-0	38 24 24	0	1.1666666666667 0 0	1.2750040305068
-0	39 24 24	0	1.25 0 0	1.2550856641843
-0	40 24 24	0	1.3333333333333 0 0	1.2378493502812
-0	41 24 24	0	1.4166666666667 0 0	1.2227888182481
-0	42 24 24	0	1.5 0 0	1.209517428715
-0	43 24 24	0	1.5833333333333 0 0	1.1977348653292
-0	44 24 24	0	1.6666666666667 0 0	1.1872043530817
-0	45 24 24	0	1.75 0 0	1.1777367206214
-0	46 24 24	0	1.8333333333333 0 0	1.1691790248092
-0	47 24 24	0	1.9166666666667 0 0	1.1614062848245
-0	48 24 24	0	2 0 0	1.1543153783021
+0	0 24 24	0	-2 0 0	1.14624279482
+0	1 24 24	0	-1.9166666666667 0 0	1.152944872139
+0	2 24 24	0	-1.8333333333333 0 0	1.1602897027505
+0	3 24 24	0	-1.75 0 0	1.168374246155
+0	4 24 24	0	-1.6666666666667 0 0	1.1773159611469
+0	5 24 24	0	-1.5833333333333 0 0	1.1872585155195
+0	6 24 24	0	-1.5 0 0	1.1983795130453
+0	7 24 24	0	-1.4166666666667 0 0	1.2109011195651
+0	8 24 24	0	-1.3333333333333 0 0	1.2251049389755
+0	9 24 24	0	-1.25 0 0	1.2413532587527
+0	10 24 24	0	-1.1666666666667 0 0	1.2601200822728
+0	11 24 24	0	-1.0833333333333 0 0	1.2820376279479
+0	12 24 24	0	-1 0 0	1.3079680700451
+0	13 24 24	0	-0.91666666666667 0 0	1.3391180280663
+0	14 24 24	0	-0.83333333333333 0 0	1.3772286479624
+0	15 24 24	0	-0.75 0 0	1.424599531194
+0	16 24 24	0	-0.66666666666667 0 0	1.4822523072674
+0	17 24 24	0	-0.58333333333333 0 0	1.550046684376
+0	18 24 24	0	-0.5 0 0	1.6269500837445
+0	19 24 24	0	-0.41666666666667 0 0	1.7105254896411
+0	20 24 24	0	-0.33333333333333 0 0	1.7963752952565
+0	21 24 24	0	-0.25 0 0	1.8777980305223
+0	22 24 24	0	-0.16666666666667 0 0	1.9461040810892
+0	23 24 24	0	-0.083333333333333 0 0	1.9920417387942
+0	24 24 24	0	0 0 0	2.008290894657
+0	25 24 24	0	0.083333333333333 0 0	1.9920417387942
+0	26 24 24	0	0.16666666666667 0 0	1.9461040810892
+0	27 24 24	0	0.25 0 0	1.8777980305223
+0	28 24 24	0	0.33333333333333 0 0	1.7963752952565
+0	29 24 24	0	0.41666666666667 0 0	1.7105254896411
+0	30 24 24	0	0.5 0 0	1.6269500837445
+0	31 24 24	0	0.58333333333333 0 0	1.550046684376
+0	32 24 24	0	0.66666666666667 0 0	1.4822523072674
+0	33 24 24	0	0.75 0 0	1.424599531194
+0	34 24 24	0	0.83333333333333 0 0	1.3772286479624
+0	35 24 24	0	0.91666666666667 0 0	1.3391180280663
+0	36 24 24	0	1 0 0	1.3079680700451
+0	37 24 24	0	1.0833333333333 0 0	1.2820376279479
+0	38 24 24	0	1.1666666666667 0 0	1.2601200822728
+0	39 24 24	0	1.25 0 0	1.2413532587527
+0	40 24 24	0	1.3333333333333 0 0	1.2251049389755
+0	41 24 24	0	1.4166666666667 0 0	1.2109011195651
+0	42 24 24	0	1.5 0 0	1.1983795130453
+0	43 24 24	0	1.5833333333333 0 0	1.1872585155195
+0	44 24 24	0	1.6666666666667 0 0	1.1773159611469
+0	45 24 24	0	1.75 0 0	1.168374246155
+0	46 24 24	0	1.8333333333333 0 0	1.1602897027505
+0	47 24 24	0	1.9166666666667 0 0	1.152944872139
+0	48 24 24	0	2 0 0	1.14624279482
