+++ TOVola/test/Simple_test/press.x.asc
@@ -1,52 +1,52 @@
 # 1D ASCII output created by CarpetIOASCII
 #
-0	0 24 24	0	-2 0 0	1.0010844469964e-26
+0	0 24 24	0	-2 0 0	1e-26
 0	1 24 24	0	-1.9166666666667 0 0	1e-26
 0	2 24 24	0	-1.8333333333333 0 0	1e-26
 0	3 24 24	0	-1.75 0 0	1e-26
 0	4 24 24	0	-1.6666666666667 0 0	1e-26
 0	5 24 24	0	-1.5833333333333 0 0	1e-26
-0	6 24 24	0	-1.5 0 0	1e-26
-0	7 24 24	0	-1.4166666666667 0 0	1e-26
+0	6 24 24	0	-1.5 0 0	1.0023466244447e-26
+0	7 24 24	0	-1.4166666666667 0 0	1.0010844469964e-26
 0	8 24 24	0	-1.3333333333333 0 0	1e-26
 0	9 24 24	0	-1.25 0 0	1e-26
-0	10 24 24	0	-1.1666666666667 0 0	1e-26
-0	11 24 24	0	-1.0833333333333 0 0	1e-26
+0	10 24 24	0	-1.1666666666667 0 0	1.0023466244447e-26
+0	11 24 24	0	-1.0833333333333 0 0	1.0010844469964e-26
 0	12 24 24	0	-1 0 0	1e-26
-0	13 24 24	0	-0.91666666666667 0 0	1.0029777131689e-26
-0	14 24 24	0	-0.83333333333333 0 0	1e-26
-0	15 24 24	0	-0.75 0 0	1.3737826154922e-05
-0	16 24 24	0	-0.66666666666667 0 0	0.00025381551325947
-0	17 24 24	0	-0.58333333333333 0 0	0.00094626251966999
-0	18 24 24	0	-0.5 0 0	0.0023225479836807
-0	19 24 24	0	-0.41666666666667 0 0	0.0046126258216895
-0	20 24 24	0	-0.33333333333333 0 0	0.0079379587573838
-0	21 24 24	0	-0.25 0 0	0.012127998740892
-0	22 24 24	0	-0.16666666666667 0 0	0.016525294300887
-0	23 24 24	0	-0.083333333333333 0 0	0.019986868382284
-0	24 24 24	0	0 0 0	0.021315886477851
-0	25 24 24	0	0.083333333333333 0 0	0.019986868382284
-0	26 24 24	0	0.16666666666667 0 0	0.016525294300887
-0	27 24 24	0	0.25 0 0	0.012127998740892
-0	28 24 24	0	0.33333333333333 0 0	0.0079379587573839
-0	29 24 24	0	0.41666666666667 0 0	0.0046126258216896
-0	30 24 24	0	0.5 0 0	0.0023225479836807
-0	31 24 24	0	0.58333333333333 0 0	0.00094626251966999
-0	32 24 24	0	0.66666666666667 0 0	0.00025381551325948
-0	33 24 24	0	0.75 0 0	1.3737826154922e-05
+0	13 24 24	0	-0.91666666666667 0 0	1e-26
+0	14 24 24	0	-0.83333333333333 0 0	1.0036088018931e-26
+0	15 24 24	0	-0.75 0 0	6.097510988301e-05
+0	16 24 24	0	-0.66666666666667 0 0	0.00036667141715058
+0	17 24 24	0	-0.58333333333333 0 0	0.0010567075322008
+0	18 24 24	0	-0.5 0 0	0.002277998915698
+0	19 24 24	0	-0.41666666666667 0 0	0.004145182320746
+0	20 24 24	0	-0.33333333333333 0 0	0.006664863677866
+0	21 24 24	0	-0.25 0 0	0.0096356893254129
+0	22 24 24	0	-0.16666666666667 0 0	0.012578136395646
+0	23 24 24	0	-0.083333333333333 0 0	0.014794428272105
+0	24 24 24	0	0 0 0	0.015624991680738
+0	25 24 24	0	0.083333333333333 0 0	0.014794428272105
+0	26 24 24	0	0.16666666666667 0 0	0.012578136395646
+0	27 24 24	0	0.25 0 0	0.0096356893254129
+0	28 24 24	0	0.33333333333333 0 0	0.0066648636778661
+0	29 24 24	0	0.41666666666667 0 0	0.0041451823207461
+0	30 24 24	0	0.5 0 0	0.002277998915698
+0	31 24 24	0	0.58333333333333 0 0	0.0010567075322007
+0	32 24 24	0	0.66666666666667 0 0	0.00036667141715056
+0	33 24 24	0	0.75 0 0	6.097510988301e-05
 0	34 24 24	0	0.83333333333333 0 0	1e-26
 0	35 24 24	0	0.91666666666667 0 0	1e-26
 0	36 24 24	0	1 0 0	1e-26
-0	37 24 24	0	1.0833333333333 0 0	1e-26
-0	38 24 24	0	1.1666666666667 0 0	1e-26
+0	37 24 24	0	1.0833333333333 0 0	1.0010844469964e-26
+0	38 24 24	0	1.1666666666667 0 0	1.0023466244447e-26
 0	39 24 24	0	1.25 0 0	1e-26
 0	40 24 24	0	1.3333333333333 0 0	1e-26
-0	41 24 24	0	1.4166666666667 0 0	1e-26
-0	42 24 24	0	1.5 0 0	1e-26
+0	41 24 24	0	1.4166666666667 0 0	1.0010844469964e-26
+0	42 24 24	0	1.5 0 0	1.0023466244447e-26
 0	43 24 24	0	1.5833333333333 0 0	1e-26
 0	44 24 24	0	1.6666666666667 0 0	1e-26
 0	45 24 24	0	1.75 0 0	1e-26
 0	46 24 24	0	1.8333333333333 0 0	1e-26
 0	47 24 24	0	1.9166666666667 0 0	1e-26
-0	48 24 24	0	2 0 0	1.0010844469964e-26
+0	48 24 24	0	2 0 0	1e-26
