+++ TOVola/test/Simple_test/rho.x.asc
@@ -15,25 +15,25 @@
 0	12 24 24	0	-1 0 0	1e-13
 0	13 24 24	0	-0.91666666666667 0 0	1e-13
 0	14 24 24	0	-0.83333333333333 0 0	1e-13
-0	15 24 24	0	-0.75 0 0	0.0037064573591128
-0	16 24 24	0	-0.66666666666667 0 0	0.015931588535343
-0	17 24 24	0	-0.58333333333333 0 0	0.030761380327774
-0	18 24 24	0	-0.5 0 0	0.048192820872831
-0	19 24 24	0	-0.41666666666667 0 0	0.067916314841793
-0	20 24 24	0	-0.33333333333333 0 0	0.08909522297735
-0	21 24 24	0	-0.25 0 0	0.11012719346688
-0	22 24 24	0	-0.16666666666667 0 0	0.12855074601451
-0	23 24 24	0	-0.083333333333333 0 0	0.14137492133432
-0	24 24 24	0	0 0 0	0.145999611225
-0	25 24 24	0	0.083333333333333 0 0	0.14137492133432
-0	26 24 24	0	0.16666666666667 0 0	0.12855074601451
-0	27 24 24	0	0.25 0 0	0.11012719346688
-0	28 24 24	0	0.33333333333333 0 0	0.089095222977351
-0	29 24 24	0	0.41666666666667 0 0	0.067916314841793
-0	30 24 24	0	0.5 0 0	0.048192820872831
-0	31 24 24	0	0.58333333333333 0 0	0.030761380327774
-0	32 24 24	0	0.66666666666667 0 0	0.015931588535343
-0	33 24 24	0	0.75 0 0	0.0037064573591128
+0	15 24 24	0	-0.75 0 0	0.0078086560861266
+0	16 24 24	0	-0.66666666666667 0 0	0.01914866619769
+0	17 24 24	0	-0.58333333333333 0 0	0.032507038194839
+0	18 24 24	0	-0.5 0 0	0.047728386896039
+0	19 24 24	0	-0.41666666666667 0 0	0.064383090332369
+0	20 24 24	0	-0.33333333333333 0 0	0.081638616339732
+0	21 24 24	0	-0.25 0 0	0.098161547081395
+0	22 24 24	0	-0.16666666666667 0 0	0.11215229108514
+0	23 24 24	0	-0.083333333333333 0 0	0.12163234878973
+0	24 24 24	0	0 0 0	0.12499993344591
+0	25 24 24	0	0.083333333333333 0 0	0.12163234878972
+0	26 24 24	0	0.16666666666667 0 0	0.11215229108514
+0	27 24 24	0	0.25 0 0	0.098161547081395
+0	28 24 24	0	0.33333333333333 0 0	0.081638616339732
+0	29 24 24	0	0.41666666666667 0 0	0.064383090332369
+0	30 24 24	0	0.5 0 0	0.047728386896039
+0	31 24 24	0	0.58333333333333 0 0	0.032507038194839
+0	32 24 24	0	0.66666666666667 0 0	0.01914866619769
+0	33 24 24	0	0.75 0 0	0.0078086560861266
 0	34 24 24	0	0.83333333333333 0 0	1e-13
 0	35 24 24	0	0.91666666666667 0 0	1e-13
 0	36 24 24	0	1 0 0	1e-13
